Respect cloze settings
