+++ src/main.ts
@@ -25,6 +25,7 @@ import { ReviewDeck, ReviewDeckSelectionModal } from "src/review-deck";
 import { t } from "src/lang/helpers";
 import { parse } from "src/parser";
 import { appIcon } from "src/icons/appicon";
+import { matchClozesWithinCardText } from "./cloze-matching";
 
 interface PluginData {
     settings: SRSettings;
@@ -690,33 +691,16 @@ export default class SRPlugin extends Plugin {
 
             const siblingMatches: [string, string][] = [];
             if (cardType === CardType.Cloze) {
-                const siblings: RegExpMatchArray[] = [];
-                if (settings.convertHighlightsToClozes) {
-                    siblings.push(...cardText.matchAll(/==(.*?)==/gm));
-                }
-                if (settings.convertBoldTextToClozes) {
-                    siblings.push(...cardText.matchAll(/\*\*(.*?)\*\*/gm));
-                }
-                if (settings.convertCurlyBracketsToClozes) {
-                    siblings.push(...cardText.matchAll(/{{(.*?)}}/gm));
-                }
-                siblings.sort((a, b) => {
-                    if (a.index < b.index) {
-                        return -1;
-                    }
-                    if (a.index > b.index) {
-                        return 1;
-                    }
-                    return 0;
-                });
+                const front = matchClozesWithinCardText(cardText, this.data.settings).reduce(
+                    (acc, sibling) => {
+                        const inputHTML = `<input type="text" class="cloze-input" size="${sibling[1].length}" />`;
 
-                const front = siblings.reduce((acc, sibling) => {
-                    const inputHTML = `<input type="text" class="cloze-input" size="${sibling[1].length}" />`;
-
-                    return acc
-                        ? acc.replace(sibling[0], inputHTML)
-                        : acc + sibling.input.replace(sibling[0], inputHTML);
-                }, "");
+                        return acc
+                            ? acc.replace(sibling[0], inputHTML)
+                            : acc + sibling.input.replace(sibling[0], inputHTML);
+                    },
+                    ""
+                );
 
                 // back is being created in flashcard-modal.tsx with getClozeBackView()
                 siblingMatches.push([front, ""]);
